+++ changelog.md
@@ -4,6 +4,7 @@
 
 -   Fixed package main file in package.json
 -   Fixed directory creation if already exists.
+-   Versioned built commands.ts.
 
 ### 0.0.3